@@ -2,20 +2,18 @@ package com.xwray.groupie.example.databinding;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
-
-import androidx.annotation.NonNull;
-import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.os.Handler;
-import androidx.core.content.ContextCompat;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.ItemTouchHelper;
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.Toast;
-
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+import androidx.databinding.DataBindingUtil;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
 import com.xwray.groupie.ExpandableGroup;
 import com.xwray.groupie.Group;
 import com.xwray.groupie.GroupieAdapter;
@@ -27,6 +25,7 @@ import com.xwray.groupie.Section;
 import com.xwray.groupie.TouchCallback;
 import com.xwray.groupie.example.core.InfiniteScrollListener;
 import com.xwray.groupie.example.core.Prefs;
+import com.xwray.groupie.example.core.R;
 import com.xwray.groupie.example.core.SettingsActivity;
 import com.xwray.groupie.example.core.decoration.CarouselItemDecoration;
 import com.xwray.groupie.example.core.decoration.DebugItemDecoration;
@@ -42,15 +41,14 @@ import com.xwray.groupie.example.databinding.item.HeartCardItem;
 import com.xwray.groupie.example.databinding.item.SmallCardItem;
 import com.xwray.groupie.example.databinding.item.SwipeToDeleteItem;
 import com.xwray.groupie.example.databinding.item.UpdatableItem;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import org.jetbrains.annotations.Contract;
 
 public class MainActivity extends AppCompatActivity {
 
     public static final String INSET_TYPE_KEY = "inset_type";
-    public static final String FULL_BLEED = "full_bleed";
     public static final String INSET = "inset";
 
     private ActivityMainBinding binding;
@@ -77,7 +75,10 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_main);
+        binding = DataBindingUtil.setContentView(
+                this,
+                com.xwray.groupie.example.databinding.R.layout.activity_main
+        );
         prefs = Prefs.get(this);
 
         gray = ContextCompat.getColor(this, R.color.background);
@@ -100,7 +101,8 @@ public class MainActivity extends AppCompatActivity {
         recyclerView.addItemDecoration(new DebugItemDecoration(this));
         recyclerView.setAdapter(groupAdapter);
         recyclerView.addOnScrollListener(new InfiniteScrollListener(layoutManager) {
-            @Override public void onLoadMore(int currentPage) {
+            @Override
+            public void onLoadMore(int currentPage) {
                 for (int i = 0; i < 5; i++) {
                     infiniteLoadingSection.add(new CardItem());
                 }
@@ -110,11 +112,9 @@ public class MainActivity extends AppCompatActivity {
         ItemTouchHelper itemTouchHelper = new ItemTouchHelper(touchCallback);
         itemTouchHelper.attachToRecyclerView(recyclerView);
 
-        binding.fab.setOnClickListener(new View.OnClickListener() {
-            @Override public void onClick(View view) {
-                startActivity(new Intent(MainActivity.this, SettingsActivity.class));
-            }
-        });
+        binding.fab.setOnClickListener(view ->
+                startActivity(new Intent(MainActivity.this, SettingsActivity.class))
+        );
 
         prefs.registerListener(onSharedPrefChangeListener);
     }
@@ -128,25 +128,20 @@ public class MainActivity extends AppCompatActivity {
 
         // Update in place group
         Section updatingSection = new Section();
-        View.OnClickListener onShuffleClicked = new View.OnClickListener() {
-            @Override public void onClick(View view) {
-                List<UpdatableItem> shuffled = new ArrayList<>(updatableItems);
-                Collections.shuffle(shuffled);
-                updatingGroup.update(shuffled);
+        View.OnClickListener onShuffleClicked = view -> {
+            List<UpdatableItem> shuffled = new ArrayList<>(updatableItems);
+            Collections.shuffle(shuffled);
+            updatingGroup.update(shuffled);
 
-                // You can also do this by forcing a change with payload
-                binding.recyclerView.post(new Runnable() {
-                    @Override public void run() {
-                        binding.recyclerView.invalidateItemDecorations();
-                    }
-                });
-            }
+            // You can also do this by forcing a change with payload
+            binding.recyclerView.post(() -> binding.recyclerView.invalidateItemDecorations());
         };
         HeaderItem updatingHeader = new HeaderItem(
                 R.string.updating_group,
                 R.string.updating_group_subtitle,
                 R.drawable.shuffle,
-                onShuffleClicked);
+                onShuffleClicked
+        );
         updatingSection.setHeader(updatingHeader);
         updatingGroup = new Section();
         updatableItems = new ArrayList<>();
@@ -158,7 +153,10 @@ public class MainActivity extends AppCompatActivity {
         groupAdapter.add(updatingSection);
 
         // Expandable group
-        ExpandableHeaderItem expandableHeaderItem = new ExpandableHeaderItem(R.string.expanding_group, R.string.expanding_group_subtitle);
+        ExpandableHeaderItem expandableHeaderItem = new ExpandableHeaderItem(
+                R.string.expanding_group,
+                R.string.expanding_group_subtitle
+        );
         ExpandableGroup expandableGroup = new ExpandableGroup(expandableHeaderItem);
         for (int i = 0; i < 2; i++) {
             expandableGroup.add(new CardItem());
@@ -193,14 +191,18 @@ public class MainActivity extends AppCompatActivity {
         groupAdapter.add(dragSection);
 
         // Horizontal carousel
-        Section carouselSection = new Section(new HeaderItem(R.string.carousel, R.string.carousel_subtitle));
+        Section carouselSection = new Section(
+                new HeaderItem(R.string.carousel, R.string.carousel_subtitle)
+        );
         carouselSection.setHideWhenEmpty(true);
         Group carousel = makeCarouselGroup();
         carouselSection.add(carousel);
         groupAdapter.add(carouselSection);
 
         // Update with payload
-        Section updateWithPayloadSection = new Section(new HeaderItem(R.string.update_with_payload, R.string.update_with_payload_subtitle));
+        Section updateWithPayloadSection = new Section(
+                new HeaderItem(R.string.update_with_payload, R.string.update_with_payload_subtitle)
+        );
         for (int i = 0; i < rainbow500.length; i++) {
             updateWithPayloadSection.add(new HeartCardItem(i, onFavoriteListener));
 
@@ -212,6 +214,8 @@ public class MainActivity extends AppCompatActivity {
         groupAdapter.add(infiniteLoadingSection);
     }
 
+    @NonNull
+    @Contract(" -> new")
     private ColumnGroup makeColumnGroup() {
         List<ColumnItem> columnItems = new ArrayList<>();
         for (int i = 1; i <= 5; i++) {
@@ -225,8 +229,12 @@ public class MainActivity extends AppCompatActivity {
         return new ColumnGroup(columnItems);
     }
 
+    @NonNull
+    @Contract(" -> new")
     private Group makeCarouselGroup() {
-        CarouselItemDecoration carouselDecoration = new CarouselItemDecoration(gray, betweenPadding);
+        CarouselItemDecoration carouselDecoration = new CarouselItemDecoration(
+                gray, betweenPadding
+        );
         GroupieAdapter carouselAdapter = new GroupieAdapter();
         for (int i = 0; i < 10; i++) {
             carouselAdapter.add(new CarouselCardItem(rainbow200[i]));
@@ -234,41 +242,53 @@ public class MainActivity extends AppCompatActivity {
         return new CarouselGroup(carouselDecoration, carouselAdapter);
     }
 
-    private OnItemClickListener onItemClickListener = new OnItemClickListener() {
-        @Override
-        public void onItemClick(@NonNull Item item, @NonNull View view) {
-            if (item instanceof CardItem) {
-                CardItem cardItem = (CardItem) item;
-                if (!TextUtils.isEmpty(cardItem.getText())) {
-                    Toast.makeText(MainActivity.this, cardItem.getText(), Toast.LENGTH_SHORT).show();
-                }
+    private final OnItemClickListener onItemClickListener = (item, view) -> {
+        if (item instanceof CardItem) {
+            CardItem cardItem = (CardItem) item;
+            if (!TextUtils.isEmpty(cardItem.getText())) {
+                Toast.makeText(
+                        MainActivity.this,
+                        cardItem.getText(),
+                        Toast.LENGTH_SHORT
+                ).show();
             }
         }
     };
 
-    private OnItemLongClickListener onItemLongClickListener = new OnItemLongClickListener() {
-        @Override
-        public boolean onItemLongClick(@NonNull Item item, @NonNull View view) {
-            if (item instanceof CardItem) {
-                CardItem cardItem = (CardItem) item;
-                if (!TextUtils.isEmpty(cardItem.getText())) {
-                    Toast.makeText(MainActivity.this, "Long clicked: " + cardItem.getText(), Toast.LENGTH_SHORT).show();
-                    return true;
-                }
+    private final OnItemLongClickListener onItemLongClickListener = (item, view) -> {
+        if (item instanceof CardItem) {
+            CardItem cardItem = (CardItem) item;
+            if (!TextUtils.isEmpty(cardItem.getText())) {
+                Toast.makeText(
+                        MainActivity.this,
+                        "Long clicked: " + cardItem.getText(),
+                        Toast.LENGTH_SHORT
+                ).show();
+                return true;
             }
-            return false;
         }
+        return false;
     };
 
-    @Override protected void onDestroy() {
+    @Override
+    protected void onDestroy() {
         prefs.unregisterListener(onSharedPrefChangeListener);
         super.onDestroy();
     }
 
-    private TouchCallback touchCallback = new SwipeTouchCallback() {
-        @Override public boolean onMove(@NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
-            Item item = groupAdapter.getItem(viewHolder.getBindingAdapterPosition());
-            Item targetItem = groupAdapter.getItem(target.getBindingAdapterPosition());
+    private final TouchCallback touchCallback = new SwipeTouchCallback() {
+        @Override
+        public boolean onMove(
+                @NonNull RecyclerView recyclerView,
+                @NonNull RecyclerView.ViewHolder viewHolder,
+                @NonNull RecyclerView.ViewHolder target
+        ) {
+            Item<GroupieViewHolder> item = groupAdapter.getItem(
+                    viewHolder.getBindingAdapterPosition()
+            );
+            Item<GroupieViewHolder> targetItem = groupAdapter.getItem(
+                    target.getBindingAdapterPosition()
+            );
 
             List<Group> dragItems = dragSection.getGroups();
             int targetIndex = dragItems.indexOf(targetItem);
@@ -288,36 +308,36 @@ public class MainActivity extends AppCompatActivity {
             return true;
         }
 
-        @Override public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
-            Item item = groupAdapter.getItem(viewHolder.getBindingAdapterPosition());
+        @Override
+        public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
+            Item<GroupieViewHolder> item = groupAdapter.getItem(
+                    viewHolder.getBindingAdapterPosition()
+            );
             // Change notification to the adapter happens automatically when the section is
             // changed.
             swipeSection.remove(item);
         }
     };
 
-    private SharedPreferences.OnSharedPreferenceChangeListener onSharedPrefChangeListener =
+    private final SharedPreferences.OnSharedPreferenceChangeListener onSharedPrefChangeListener =
             new SharedPreferences.OnSharedPreferenceChangeListener() {
-                @Override public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
+                @Override
+                public void onSharedPreferenceChanged(
+                        SharedPreferences sharedPreferences,
+                        String s
+                ) {
                     // This is pretty evil, try not to do this
                     groupAdapter.notifyDataSetChanged();
                 }
             };
 
-    private Handler handler = new Handler();
-    private HeartCardItem.OnFavoriteListener onFavoriteListener = new HeartCardItem.OnFavoriteListener() {
-        @Override
-        public void onFavorite(final HeartCardItem item, final boolean favorite) {
-            // Pretend to make a network request
-            handler.postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    // Network request was successful!
-                    item.setFavorite(favorite);
-                    item.notifyChanged(HeartCardItem.FAVORITE);
-                }
-            }, 1000);
-        }
+    private final Handler handler = new Handler();
+    private final HeartCardItem.OnFavoriteListener onFavoriteListener = (item, favorite) -> {
+        // Pretend to make a network request
+        handler.postDelayed(() -> {
+            // Network request was successful!
+            item.setFavorite(favorite);
+            item.notifyChanged(HeartCardItem.FAVORITE);
+        }, 1000);
     };
-
 }
